requirements.txt
asyncssh>=2.14.0
mcp[cli]>=1.0.0
pyyaml>=6.0.1
uvicorn[standard]>=0.29.0
starlette>=0.37.0
